Add primitive MinIO client
We plan to add a more comprehensive MinIO client, that uses mc or HTTP endpoints,
but it will take some time until it is ready. To unblock PostgreSQL backup/restore,
I am adding a very primitive Minio client based on S3 client. We could continue
using the S3 client, as S3 client is compatible with MinIO. However S3 client
is only able to do S3 operations like creating buckets or listing objects, but
we also need to do things like creating new access tokens, which is in the
realm of IAM, whose client is not compatible with MinIO.