Make create_server_certificate more similar to create_root_certificate
create_root_certificate creates the root certificate and returns to the caller.
It is caller's responsibility to save it to the database. On the other hand
create_server_certificate both creates the certificate and save it into the
database. It also pushes the certificate to the backing VM. To remove this
discrepancy, we changed create_server_certificate to be more similar to
create_root_certificate. This seems like causing duplication of code while
pushing the certificate to the VM, but soon, that part will be refactored
in another commit when we introduce entity managing "physical" servers.