Add PostgresServer to represent servers grouped under PostgresResource
Not to be confused with the original PostgresServer entity that is renamed as
PostgresResource. PostgresResource represents the logical entity. It is what
users see and interact. Underlying vms can change as a result of failover but
the same PostgresResource stays till user deprovisions it. Things like billing,
dns records, certificates are tied to the PostgresResource.

On the other hand, PostgresServer represents the "physical" entity in the loose
meaning of the word physical. It is ties the underlying vm to PostgresResource,
and every time the underlying vm changes, PostgresServer also changes with it.
Currently there is one PostgresServer for each PostgresResource, however when
we introduce HA standbys or scale up/down with failover there can be multiple
PostgresServers associated with given PostgresResource.

I initially considered directly using Vm entity for the same purpose as there
is one to one mapping between Vm and PostgresServer. However that turned out to
be more cumbersome that I thought. Firstly, there are some data that we want to
keep for each PostgresServer such as timeline_access (which is a concept to be
introduced in backup/restore). Such data cannot be stored in Vm as it is not
a general concept that affects all Vms. It is also not ideal to store it in the
PostgresResource because it needs to be stored in array whose ordering is not
consistent. This is not a huge problem but annoyance of course. However it is
not the only one. Without its own entity, we need to use PostgresResource to
manage these, which would make it pretty difficult to do two unrelated things
for each sub resource at the same time.