Use IP address as hostname if Postgres' DNS zone is not configured
hostname depends on the DNS zone and many things like provisioning, certificate
generation, streaming replication depends on hostname. Thus if DNS zone is not
configured (which is likely in development environments), PostgreSQL service
breaks. With this commit, we fallback to IP address in hostname, which is good
enough for almost all purposes.