Fix pagination issue when retrieving all runners for the repository
When a runner with the same name already exists, we first delete the
runner before adding it again. To do this, we need the runner_id. We
retrieve all runners for the repository and locate the specific runner
by name.

The runners endpoint returns a paginated result with a maximum per_page
value of 100. We must paginate across all pages. The "octokit.rb"
library has a `paginate` helper, and we already use it.

https://github.com/ubicloud/ubicloud/blob/d0bec18866fb5ad8c43e9c4a3f9215182fb834bf/prog/vm/github_runner.rb#L184

But it only paginates if `auto_paginate` is set to true.

https://github.com/octokit/octokit.rb/blob/ed692187cf2b8238b820e3e85a517357c42a75e3/lib/octokit/connection.rb#L86C11-L86C11