Redact sensitive and long columns from outputs
Even if some columns are encrypted, it's not advisable to print them in
logs or inspect results. Additionally, some column values are
excessively long, causing the output to be bloated.

The model redacts encrypted columns by default. If the model has
additional columns to redact, it can overwrite the existing method as
follows:

    def self.redacted_columns
        super + [:root_cert_1, :root_cert_2, :server_cert]
    end