Fix reap provoking lease violation
Many days after 20bd3edfcda39aedd7499668b23bf2aa4a57fc00, we recorded
a single lease violation message, before, they used to be frequent.

Based on the circumstances, I think the bug mechanism is like this,
and it is addressed by this patch:

    Thread A               Thread B
    run child strand
    take lease; commit
    set exitval; commit
                           read exitval; reap
    release lease; crash

This patch elects to not reap a child if it is still leased, which can
be momentarily true between committing exitval and clearing the lease.