Adds basic firewall rules for GitHub runners
With this commit, we are adding basic firewall rule support for github
actions runners. The model is very simply and the main objective of this
commit is to have an initial implementation to work on for further
development. For now, this implementation adds a basic nftables table
for firewall rules which denies incomming traffic by default. We
make use of that table by adding a rule to allow all ip addresses
(0.0.0.0/0, ::/0) for regular VMs and we delete these default rules to
block all ip addresses for github runners.

We chose to allow tcp dport 22 by default because that is the ssh port
and we have to be able to continue opening ssh connections.

We chose to allow all communication in the private network.

We also chose to allow all outgoing communication in a stateful way.

We changed the NAT logic because the previous one was not tracking the
connections which causes the logic to allow ssh connections to fail.
Simply changing saddr or daddr causes the conntrack object to not be
tracked properly for the ssh connections. Therefore, it did not work.
With this current logic, we handle the DNAT and SNAT properly and in a
tracked way which allows us to implement the tcp dport 22 allow.

For now, this commit only works for ingress.

The FirewallRules are added directly to the PrivateSubnet entity. It is
possible that in future we will have an independent firewall entity
which will have the rules as part of it. For now, that's harmless.