Fix in-memory data structure while reaping
In production this mistake in modifying the `children` list caused
`leaf?` to not function properly, returning `false` when it would have
been beneficial to return `true`.

It wasn't visible immediately, because eventually progress would be
made, when a `run(seconds)` session ended and the entire strand got
reloaded.  But instead of taking a few milliseconds, it would take
`seconds` time, spinning unproductive logs all the while.  Eventually
enough Strands got into this state, causing intermittent failures in
the heartbeat monitoring that tipped us off to an abnormality.

Reviewing log throughput of an affected prog label clearly showed the
problem and its duration correlating with the deployment of
44316cef651e2169f0556924f6d96e74063049e5.

The test included fails when the fix to `reap` is likewise removed.