Learn CPU dies out of Linux CPU topology
Previously, in the ancient era (April, in
d30c157f2f1b2f3d7146824b836da7f4d6aa3973), I used NUMA nodes, the main
die-correlated statistic out of `lscpu` to compute the cpu topology.
However, in practice, we wind up with far fewer NUMA nodes because
BIOS configurations tend to interleave die memory access.

But, this is wrong: we can do better.

The first step is to learn the die count, and that is what this patch
does.  The second step, not done yet, is to use that die count to
calculate arguments to `cloud-hypervisor`.

== Background

Recently, I did some kernel source code reading, and learned that dies
can be used in the scheduler to try to take advantage of cache
unification and power management, and that they were exposed in
`/sys/devices/system/cpu/`, and are not returned decisively anywhere I
can see in `lscpu`, which is how I missed it back in April.

In addition, dies are not used as an abstraction in arm64, and in fact
cloud-hypervisor rejects non-`1` settings for it on that platform:

https://github.com/cloud-hypervisor/cloud-hypervisor/commit/0fc3fad363ffe43708252c9094477a12e36e5c9f

In newer versions of Linux -- not including Ubuntu 22.04 Jammy's Linux
5.15 -- `/sys/devices/system/cpu/` will learn the `cluster` level of
core aggregation, which is yet another level of hierarchy, and this
one applies to both arm64 and x64.  As-is, cloud-hypervisor is
hardcoded to place all exposed processors in a single cluster, and it
is not part of the `topology` flag passed to cloud-hypervisor.

https://lists.gnu.org/archive/html/qemu-devel/2021-03/msg09848.html
discusses how arm64 will have a four-level hierarchy, and x64 will
have a five-level one (from `dies`).

So, it looks like this is an active area of change in the last few
years, we should expect some confusion in the software stack on how
and when to best use these fields in the near future.