Don't wait vm deletion while deleting PostgresServer
PostgresServer has foreign key to the Vm, thus Vm cannot be deleted before
PostgresServer. To be able to wait, we have two options;
- Don't create foreign key
- Create foreign key but use CASCADE SET NULL

Both are not ideal. I think it would be better to skip waiting Vm deletion.
From user perspective, the deletion would complete faster and we would handle
the deletion in background, so it ends up as a better user experience. Also
Vm deletion has its own deadline, so we would be notified. If we wait for Vm
deletion, we would get two pages.