Handle nil vm case while generating hostname
We create Vm along with PostgresServer, so it should always have a Vm. However
in the deletion path, Vm would be deleted before PostgresServer. If we need to
generate hostname at that time, we would crash, which can happen if user wants
to display PostgresResource details. With this commit we are handling that nil
vm case, by returning empty string as hostname and connection_string.