Switch CPU topology calculation from nodes to package dies
This is more accurate to represent a viable CPU topology, but it also
raises the question of whether the implementation should continue to
learn NUMA nodes: they were only being used as a stand-in for dies.

It is permissible to remove multi-die Altra tests, since no arm64 chip
defines dies.  See
https://github.com/cloud-hypervisor/cloud-hypervisor/commit/0fc3fad363ffe43708252c9094477a12e36e5c9f