Render exceptions raised out of dispatcher using Clog
Previously, it would be emitted in its default `#full_message`
representation, thus, not JSON formatted text. This made it
inconvenient to look at in a JSON-parsing log utility, and also
doesn't absorb Clog's useful behavior of including the thread name
with the log record, i.e. we had no idea of the subject of the stack
trace.

A complication seen here is to emit multiple Clog records for nested
exceptions.  The reason is, I have found that some tools can have
trouble with multiple backtraces in one record (it simply gets too
big), but generally one backtrace per record is fine.

As an aside, I found a test bug here that wasn't apparent until I
added a test for this particular issue: apoptosis threads could break
the Thread-join invariant seen in the `after` block.  It just so
happened all the other tests interacted with the Ruby thread scheduler
such that it wasn't a problem.