Stop early incrementing update_firewall_rules for runner VMs
We used to increment the `update_firewall_rules` semaphore immediately
when we create the new VMs after deleting the firewall_rules, but this
would cause a crash like this, and halt progress:

    stderr: Cannot open network namespace "vm0tcpph": No such file
    or directory.

A fix for this is to not increment `update_firewall_rules`.

As long as we clear the firewall_rules that allow all traffic, we
should be good, because the firewall rules are processed when the VM
becomes ready in the NicNexus.