Do not charge  if the GitHub runner did not pick a job
When a GitHub workflow job is cancelled before being assigned to a
runner, we cannot destroy the runner upon receiving a webhook event
request, as we have not yet paired the runner ID with the job ID. To
address this issue of leftover runners, we destroy them if they do not
receive a job within 2 minutes. We should not charge the customer for
these 2 minutes.