Do not try to take backup if blob storage is not configured
For development environments, it is quite likely that we don't have all the
dependencies of PostgreSQL service ready. In such cases, it is important to
fail gracefully. We already did this for DnsZone dependency. With this commit
we are also doing that for blob storage dependency. If blob storage is not
configured, everything except backup/restore would continue to work. Of course
many other things such as HA, scale up/down etc. depends on backup/restore.
Only remaining serious feature is provisioning, which is good enough for most
development needs.