Perform user_id check before accessing project details and resources
Primary motication of this is preventing leaking information about the project
to users who are not part of it. For example, resource endpoints themselves
might want to access project's feature flags, which should only be done after
we ensure user is part of the project.

As a side benefit, this also reduces the impact if we forget about checking
permissions for a particular resource. Instead of making that resource visible
to everyone, this limits impact to the project members.