Change default gateway not to be network address for ubuntu2004
Looks like the behaviour for default gateway changed from linux kernel
5.4 to 5.15. In the new versions of the Kernel, we can use the network
address of a subnet as the default gateway in the VMs. Our DHCP system
sends the gateway from the route (the ip address in the tap device).
With 5.4 kernel, if we send the network address as the gateway, dhcp
procedure fails to add the default route. Therefore, we are making a
small change and using the next address of the network address as a
gateway.