Change the account subpages navigator from tab bar to sidebar
Our "My Account" page has several subpages such as  "Change Password",
"Close Account". Currently, we use a tab bar to switch between these
subpages. However, this component isn't suitable when the number of
subpages increases. I'm working on the 2FA feature, which will add more
subpages. Additionally, the tab bar's responsiveness is poor, leading to
display issues on smaller screens.

Previously, we used a side menu component from Tailwind UI for the
project's subpages, but removed it when we transitioned to project-based
views.

In my opinion, the side menu is superior to the tab bar.