Allow to log additional fields with models in clog
Sometimes, the model data is insufficient, and we may need to add extra
fields to the clog's metadata. The new array metadata feature achieves
this.

When the metadata is an array, it parses each element sequentially. If
the element is a Sequel::Model, it obtains its clog representation,
similar to the previous commit. If the element is a hash, it merges with
the intermediate hash.

For example, `[strand, {field1: "a", field2: "b"}]` expands to `{strand:
{id: 123, label: "start"}, field1: "a", field2: "b"}`.