Explicitly set StrictHostKeyChecking
Without explicitly setting these, net-ssh uses whatever default is set in the
operating system, which might be different that what we want. In sshable.cmd(),
we don't want to verify host key because IP addresses can be reassigned to the
other servers over time. In the future we might want to keep host keys in our
DB and do verification for extra security, but it is not in the scope of this
commit.