Use /dat/16/data as data directory
Default data directory in Ubuntu is /var/lib/postgresql/16/main, which we used
to use and then create symbolic link to /dat/16/data for ease of operations.
However, when we do that, we don't utilize the data disk we mounted to /dat as
the /dat directory only keeps the symbolic link file. We can mount the disk to
/var/lib/postgresql/16/main, but that is too deep in the filesystem hierarchy.
So, instead, we start to use /dat/16/data directly as data directory.