Switch to simpler up-check
The `ssh` program drags in too much possible environmental
configuration that can conflict with our intended function here.
Also, spawning a process here is wasteful.  For the purposes of this
liveness check, opening a TCP connection, then closing it without
transmitting anything, suffices.

Note that general presumption of port-22 accessibility on VMs is not
sound, but that's a problem for another day.  The current strategy did
help us identify the boot-hang seen in some Linux kernel versions with
buggy tick code, so, it deserves some credit and can live on a little
longer.