Enable multiqueue tap
While benchmarking, we have encountered bottlenecks in network queue
processing.  Scouring the web, we see that a common configuration is
to increase the number of queues, and thus the number of
cloud-hypervisor queue-pair threads, and thus the number of processors
that can be used to move ethernet frames around.

Easy enough: add some tap options, and this works if vcpus>=2.  If you
have eight vcpus, allocate 16 + 1 queues, which form eight pairs and
one control queue.

Note that this is the maximum, in the above example, allocating
eighteen queues is rejected by cloud-hypervisor complaining about
insufficient vcpus to have a point.

But, why vcpus >= 2? My patch doesn't work for arm64 with one core.
On x64 with SMT, vcpus is always two or greater.  You can simulate the
fault by making a one-vcpu x64 computer out-of-band.

The reason is an an obscure conflict with how we, uncustomary to
cloud-hypervisor, do not add the net_admin capability to the
cloud-hypervisor binary to let it manage tap devices.  Instead, we
elect to create the tap out-of-band ourselves via rhizome.  For this
to work, the tap rhizome creates must match what cloud-hypervisor
needs exactly.

This has been trivial in event queues = 3, which is the minimum and
the default: two are in a data-exchanging pair, and one is the
control.  In this case, the simplest tap device definition, the one
we've been always up through now, is always used.

But, if and only if num_queues is specified, and greater than 3, the
tap must have the vnet_hdr and multiqueue options.  This couples the
tap definition to how CPUs are configured, and rhizome is not
organized to do that.

So, I wrote this long message because it's the end of my day, but
perhaps someone can pick it up in Europe, and we can squeak out a bit
of time compression on the overall project.  To take advantage of
multiple queues, rhizome will have to be reorganized a little bit, or
we have to concede to giving cloud-hypervisor the net-admin
capability.  Although ungratfiying, the space of options on tap is
pretty short and slow-moving, so I am inclined to add complications to
create (and perhaps re-create) the taps just right.