Create a generic button component for UI
We use similar primary and delete buttons in our UI, each with a long
list of CSS classes. This redundancy becomes problematic when we want to
alter the button style, as changes need to be applied everywhere, making
it difficult to maintain. To address this, I created a generic button
component. If no link is provided, the `<button>` tag is used in place
of the `<a>` tag.