Configure OTP authentication in rodauth
OTP is one of the most popular 2FA methods. We use the `otp` plugin from
rodauth for this purpose.

The `otp_setup_*` configurations are related to OTP setup, while the
`otp_disable_*` configurations are associated with disabling OTP.

The `otp_auth_*` configurations are used for user authentication via
OTP.

I replaced 'TOPT' with 'One-Time Password' in the UI, as I believe it's
more user-friendly.