Configure WebAuthn authentication in rodauth
WebAuthn is another popular 2FA method. Security keys, such as Yubikey
and Apple passkeys, are built on WebAuthn. We use the `webauthn` plugin
from rodauth for this purpose, which also includes the necessary
JavaScript code.

The `webauthn_setup_*` configurations are related to WebAuthn key setup,
while the `webauthn_remove_*` configurations are associated with
removing webauthn key.

The `webauth_auth_*` configurations are used for user authentication via
WebAuthn keys.

I replace 'WebAuthn' texts with 'Security Keys' in the UI, as I believe
it's more user-friendly.

Also, I add the `name` property to webauthn keys. It makes the
management easier for users.