Add date validation helper
It helps us ensure that the provided date is valid. If a `Time` object
is provided instead of a string, it returns the similar object.

I added the `param` argument to raise a validation error with the
correct input name for display on the UI.

I use DateTime.parse instead of Time.parse because it uses UTC as
default timezone but Time.parse uses local timezone