Add PostgreSQL database restore functionality to the UI
We allow users to restore the database from a target date within the
range of the earliest and latest restore times. This is enforced on the
console using flatpickr's minDate/maxDate configurations. However,
merely enforcing this on the UI is insufficient, as the forced value can
be easily altered via inspect element. Thus, we also validate the
restore target on the backend.

If the server lacks an earliest_restore_time value, we do not display
the restore card at all.