Add earliest/latest restore time helpers to the pg timeline
We need to determine earliest/latest restore times to provide a specific
time range for the user on the UI.

We don't allow to restore the destroyed databases for now. So the latest
restore time is the current time.

The earliest restore time is little bit tricky. Determining the earliest
restore time is straightforward; simply query the blob storage and
select the oldest entry. However, we aim to avoid querying the blob
storage service each time a user opens the page in the UI. The
"earliest_backup_completed_at" column is used to cache the value,
eliminating the need to query the blob storage every time. The
"earliest_backup_completed_at" value can be changed when a new backup is
created or an existing backup is deleted. It's nil when the server is
created, so we get it from the blob storage until the first backup
completed. Currently, we lack a backup cleanup feature.  Once it is
implemented, we can invoke the "refresh_earliest_backup_completion_time"
method at the appropriate points.

We return earliest/latest restore times only if the server is primary.