Add --fail flag to curl  for 404 not found errors
Recently, our hourly E2E tests began failing due to "Image is not in
qcow2 format" exception. The image url was returning a 404 Not Found
error. By default, curl does not exit with failure codes for 4xx status
codes; instead, it downloads the error message as a file. Subsequently,
qemu-convert attempts to convert this error message.

I added `-f` flag to curl, forcing it to exit with a non-zero code for
4xx errors.

From curl man:

    -f, --fail

        (HTTP) Fail fast with no output at all on server errors. This is
        useful to enable scripts and users to better deal with failed
        attempts. In normal cases when an HTTP server fails to deliver a
        document, it returns an HTML document stating so (which often
        also describes why and more). This flag will prevent curl from
        outputting that and return error 22.

        This method is not fail-safe and there are occasions where
        non-successful response codes will slip through, especially when
        authentication is involved (response codes 401 and 407).