Add separate DNS record entry for each insert/delete record operation
We used to update tombstoned column when a DNS record is deleted. Records with
same name, type and data can be added and removed many times. We also tried to
keep only one copy of active record using a unique constraint. This provides
nice size reduction, but it complicates record processing. Especially when a
deleted record is added back, we don't have a way to know which request came
last (thus whether should we create or remove the record in the DNS server).

We fix this situation by adding separate DNS record entry for each insert and
delete request. We also tag each record with a timestamp. Now we can process
the records in the request order. Adding separate records for each operation
increases the total size. To remove unnecessary bloat, we expand the purge logic
to also purge obsoleted records.