Page if no backup is taken for 48 hours
We want to be notified if we cannot take backups for long duration, either as
a isolated case on single database or as a general problem in backup system.
With this change, exception we throw in PostgresTimeline when blob storage is
not configured is obsoleted, so we are removing that as well.