Fix storage_volumes value in the stack for testing
`storage_size_gib` should be `size_gib` in the frame. The current usage in
the tests is incorrect.
https://github.com/ubicloud/ubicloud/blob/d96815d8fafb12ef369e1c7cc7b0f2cd0c713edb/lib/validation.rb#L64C5-L64C5

We only use `storage_volumes` from the stack at the `start` label, after
which we clear it. Other labels should use `VmStorageVolume` records.
Consequently, I moved this general 'allow' block to the `.assemble`
block only.