MinIO Client for regular S3 operations
This part of the SDK implements regular S3 operations. As a full S3
compatible service, MinIO, our request headers should be signed
following S3's directives. It is explained in detail, here;
https://docs.aws.amazon.com/AmazonS3/latest/userguide/RESTAuthentication.html

Using the access_key, secret_key, date we sign the header, and some
other fields necessary to generate a hash in another hash, we create an
authentication string. Using that in the header, we are able to
introduce ourselves to the server for a short while (until the date is
in the range that server accepts the token).

Furthermore, this commit comes with some basic API implementation, which
are;
1. create_bucket(bucket_name)
2. delete_bucket(bucket_name)
3. bucket_exists?(bucket_name)
4. list_objects(bucket_name, folder_path, max_keys: 1000)

The only not simple one is the list_objects API since it implements
pagination on the server side which also requires us to solve it on the
client side. MinIO sends a continuation_token in the respond list that
we can use to fetch the items after the last item that we already got.
For now, we are not implementing the iterator object return but simply
collect all of the items since Postgres usecase will never return more
than 1000 backups for a specific bucket.