MinIO Client for regular MinIO Admin operations
These APIs are simple MinIO Admin operations that are available as long
as the client is someone with admin privileges. We can simply make use
of these APIs by using the admin path and sign the headers the same way
we do for any S3 request. The implemented endpoints are;
1. admin_info: gives information about the cluster in general. The
health status of nodes, drives, everything.
2. admin_policy_list: lists all of the policies created in the system
3. admin_policy_add(policy_name, policy): creates the requested policy
that is passed in a regular hash form. An example policy is the
following;
{
  "Version" => "2012-10-17",
  "Statement" => [
    {
      "Action" => ["s3:GetBucketLocation"],
      "Effect" => "Allow",
      "Principal" => {
        "AWS" => ["*"]
      },
      "Resource" => ["arn:aws:s3:::test"], "Sid" => ""
    }
  ]
}

4. admin_policy_info(policy_name): fetches information regarding the
policy stated with the policy_name.