Add storage space support
Previously there wasn't an easy way to store each disk of a VM on a
separate storage device on the host. This PR enables that by introducing
storage spaces.

The steps to use this are:
* System admin mounts a disk on the host to `/var/storage/spaces/nvme0`,
  and a different disk to `/var/storage/spaces/nvme1`.
* When creating a VM, we specify `storage_space` param of each volume:

```
> Prog::Vm::Nexus.assemble(..., storage_volumes: [{..., storage_space:
> "nvme0"}, {..., storage_space: "nvme1"}])
```

* Then, the first VM storage volume will be stored on the first physical
  disk & the second storage volume will be stored on the second physical
  disk. For example, first disk's data will go to
  `/var/storage/spaces/nvme0/vm12345/1/`

If the storage_space parameter is omitted, then the default storage
space will be used, and storage volumes will be created directly under
`/var/storage`, as before.