Change return type of list_objects for non existent bucket
When buckets do not exist, we need to gracefully return empty list.
Otherwise, this causes a failure in the postgres service provisioning
path. The main reason is, at the time of creation of the object, the
bucket for backup does not exist, yet. However, the moment we are
redirected to the postgres object page, we send a query to the MinIO
service. Just returning empty array is more than enough to fix that.