Add firewall rule ip,port_range tuples for nftables
With this commit, we add 2 more sets to the nftables configuration so
that the port ranges are usable for the firewall rules. This will be
specifically used for postgresql service since it's supposed to serve
requests from one and only 5432 port.

Here, there is one more change to make the migration efforts easier.
Instead of flushing individual sets and repopulating, we start to use
delete table command so that we can make change to the whole table
definition without changing the dataplane code. With this change, the
firewall rules related parts in the dataplane script vm_setup is also
obsolete. Therefore, we remove it.