Always skip first address of subnet
9f0fced2c3ad738db83674f214951179bfe4896a was an expedient fix to fix
Ubuntu 20.04, because sometime after that, Linux increased the
flexibility of the definition of subnets to support using the all-zero
suffix for an interface's normal address.

The problem is, in general, we won't know what version of the kernel
is booting, and to do things this way is contra-norms.  It also may
not work on some other operating systems because it is contra-norms.

We're still early enough where we can probably sneak this in without a
huge project, and I don't think using a the lowest common denominator
here is practically nor aesthetically offensive.